+++ src/anticarium_pi/anticarium_pi/i2c_members/I2CInput.cpp
@@ -9,16 +9,17 @@
 I2CInput::I2CInput(const I2CSlaveParameters& params, QObject* parent) : I2CSlave(params, parent) {
 }
 
-shared_types::SensorData I2CInput::getSensorData() {
+const shared_types::SensorData& I2CInput::getSensorData() const {
     return sensorData;
 }
 
 void I2CInput::fetchData() {
     auto buffer = std::make_unique<unsigned char[]>(params.bufferSize);
 
-    int bytesRead = read(file, buffer.get(), params.bufferSize);
+    const int bytesRead = read(file, buffer.get(), params.bufferSize);
     if (bytesRead == static_cast<int>(params.bufferSize)) // read() returns the number of bytes actually read, if it doesn't match then an error occurred (e.g. no response from the device)
     {
+        // First byte indicates data type
         auto value = *reinterpret_cast<float*>(buffer.get() + 1);
 
         switch (buffer[0]) {
@@ -29,8 +30,13 @@ void I2CInput::fetchData() {
                 sensorData.setHumidity(static_cast<int>(value));
             } break;
             case 'm': {
-                int mappedMoisture = map<int>(static_cast<int>(value), maxAnalogMoisture, minAnalogMoisture, 0, 100);
-                sensorData.setMoisture(mappedMoisture);
+                value = map<int>(static_cast<int>(value), maxAnalogMoisture, minAnalogMoisture, minPercentage, maxPercentage);
+
+                // There are cases where map outputs bigger value by 1
+                if (value > maxPercentage) {
+                    value = maxPercentage;
+                }
+                sensorData.setMoisture(static_cast<int>(value));
             } break;
             default: {
             } break;