+++ src/anticarium_pi/anticarium_pi/i2c_members/I2CInput.h
@@ -6,7 +6,7 @@ class I2CInput : public I2CSlave {
     Q_OBJECT
   public:
     I2CInput(const I2CSlaveParameters& params, QObject* parent = nullptr);
-    shared_types::SensorData getSensorData();
+    const shared_types::SensorData& getSensorData() const;
 
   public slots:
     /*
@@ -17,6 +17,6 @@ class I2CInput : public I2CSlave {
   private:
     shared_types::SensorData sensorData;
 
-    const int minAnalogMoisture = 17;
+    const int minAnalogMoisture = 0;
     const int maxAnalogMoisture = 1023;
 };